Update payout amount validation to require integer values
- Changed the validation rule for the 'amount' field in StoreRequest from 'numeric' to 'integer' to ensure only whole numbers are accepted, enhancing data integrity for payout processing.

+++ app/Http/Requests/API/Payout/StoreRequest.php
@@ -17,7 +17,7 @@ class StoreRequest extends FormRequest
     {
         return [
             'merchant_id' => ['required', 'exists:merchants,uuid'],
-            'amount' => ['required', 'numeric', 'gt:0'],
+            'amount' => ['required', 'integer', 'gt:0'],
             'payout_method_type' => ['required', 'string', Rule::in(PayoutMethodType::values())],
             'payment_method_id' => [
                 'required',
